change zoom to double instead of int, for smooth scrolling

+++ src/main/java/ch/awae/simtrack/view/IViewPort.java
@@ -82,7 +82,7 @@ public interface IViewPort {
 	 */
 	public boolean isVisible(TileCoordinate hex);
 
-	public int getZoom();
+	public double getZoom();
 
 	public Point getSceneCorner();
